Revert "Add `Triton` curve (#3)"
This reverts commit fa86a371a0290ae4c5887de0fe2f60ad22bfd3f4.
