+++ .github/workflows/ci.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       matrix:
         include:
-          - rust: 1.67.0
+          - rust: 1.63.0
             feature: default
           - rust: nightly
             feature: asm
@@ -37,7 +37,7 @@ jobs:
     strategy:
       matrix:
         include:
-          - rust: 1.67.0
+          - rust: 1.63.0
             feature: default
           - rust: nightly
             feature: asm
