+++ rust-toolchain
@@ -1,1 +1,1 @@
-1.67.0
+nightly-2022-08-23
