+++ src/pluto_eris/mod.rs
@@ -1,11 +1,8 @@
-// mod eris;
-mod curve;
+mod eris;
 mod fields;
-// mod pluto;
-// mod triton;
-mod engine;
+mod pluto;
 
-pub use curve::*;
-pub use engine::*;
+pub use eris::Eris;
 pub use fields::fp::*;
 pub use fields::fq::*;
+pub use pluto::Pluto;
